Use a baudrate that can be cleanly derived from the system clock
Standard serial baudrates like 9600 can be derived only approximately,
leading to random errors.

Most usb-serial and usb-rs485 converters can do pretty much any baudrate
just fine so a non-standard baudrate should not be much of an issue.

I tested it with a CP2101 at 25000 and it seems to work fine